Warning when creating meetings in public channels (#48)
Added a warning when creating a meeting in a public channel. Note that the user can choose to never see the warning again, and this doesn't happen in club channels. Fixes #48.

The current message looks like this:

![Screenshot of ephemeral Slack message](https://doggo.ninja/6hAvRj.png)

**Diff/review expectations:**

- Prettier decided to do some auto-formatting so that might screw up the diff a bit, but I chose not to stage files where formatting was the only change.
- Moved everything from `slash-z.js` into `slash-z-inner.js` which _isn't_ en endpoint but can be called from multiple places. This is required so we can continue the /z flow from a button interaction.
- Added endpoint for interactivity (buttons) and updated the Slack manifest. Consideration: right now all the handling code is in `interactivity.js`, it might be more in-line with code style to abstract them into functions somewhere else. Happy to do this.
- Added to the Prisma schema to keep track of users who've dismissed the warning. I'm not very used to Prisma so I'm not sure if I did the migration correctly. Open to better naming ideas for the table.

**Alternatives & considerations:**

As suggested by @cosmicoptima, it might be less intrusive to remove "don't show this again" as a secondary option and make that the default; in other words, only ever show the warning once and change the copy to be more tutorial-style.

It might be tedious to show this warning in channels like #hack-night and #productivity where meetings are frequent and intentional. Would it be a good idea to have a whitelist? The "don't show this again" feature might suffice.